# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Northwind Knowledge RAG API",
    "description": "Permission-aware RAG: collections, ingestion, hybrid search, cited chat answers, analytics and eval.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/analytics/overview": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Analytics Overview",
        "operationId": "analyticsOverview",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/analytics/unanswered": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Analytics Unanswered",
        "operationId": "analyticsUnanswered",
        "parameters": [
          {
            "name": "fresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Fresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnansweredCluster"
                  },
                  "title": "Response Analyticsunanswered"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/analytics/negative-feedback": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Analytics Negative Feedback",
        "operationId": "analyticsNegativeFeedback",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NegativeFeedbackItem"
                  },
                  "title": "Response Analyticsnegativefeedback"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/query-logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Query Logs",
        "operationId": "listQueryLogs",
        "parameters": [
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueryLogSummary"
                  },
                  "title": "Response Listquerylogs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/query-logs/{query_log_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Query Log",
        "operationId": "getQueryLog",
        "parameters": [
          {
            "name": "query_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Query Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTrace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClusterQuestion": {
        "properties": {
          "query_log_id": {
            "type": "string",
            "title": "Query Log Id"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "query_log_id",
          "question",
          "user",
          "created_at"
        ],
        "title": "ClusterQuestion"
      },
...
      "CollectionRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CollectionRef"
      },
...
      "DailyPoint": {
        "properties": {
          "day": {
            "type": "string",
            "title": "Day"
          },
          "questions": {
            "type": "integer",
            "title": "Questions"
          },
          "no_answer": {
            "type": "integer",
            "title": "No Answer"
          },
          "users": {
            "type": "integer",
            "title": "Users"
          },
          "p95_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Ms"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          }
        },
        "type": "object",
        "required": [
          "day",
          "questions",
          "no_answer",
          "users",
          "p95_ms",
          "cost_usd"
        ],
        "title": "DailyPoint"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NegativeFeedbackItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          },
          "query_log_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Log Id"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "condensed_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condensed Question"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "retrieved": {
            "items": {
              "$ref": "#/components/schemas/RetrievedCandidate"
            },
            "type": "array",
            "title": "Retrieved"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reason",
          "comment",
          "created_at",
          "user",
          "message_id",
          "answer",
          "citations",
          "query_log_id",
          "question",
          "condensed_question",
          "outcome",
          "retrieved"
        ],
        "title": "NegativeFeedbackItem"
      },
      "Overview": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          },
          "questions": {
            "type": "integer",
            "title": "Questions"
          },
          "active_users": {
            "type": "integer",
            "title": "Active Users"
          },
          "no_answer": {
            "type": "integer",
            "title": "No Answer"
          },
          "no_answer_rate": {
            "type": "number",
            "title": "No Answer Rate"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          },
          "feedback_up": {
            "type": "integer",
            "title": "Feedback Up"
          },
          "feedback_down": {
            "type": "integer",
            "title": "Feedback Down"
          },
          "latency_ms": {
            "$ref": "#/components/schemas/Percentiles"
          },
          "first_token_ms": {
            "$ref": "#/components/schemas/Percentiles"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "cost_per_question_usd": {
            "type": "number",
            "title": "Cost Per Question Usd"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenTotals"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyPoint"
            },
            "type": "array",
            "title": "Daily"
          },
          "steps_ms": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Percentiles"
            },
            "type": "object",
            "title": "Steps Ms"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "questions",
          "active_users",
          "no_answer",
          "no_answer_rate",
          "errors",
          "feedback_up",
          "feedback_down",
          "latency_ms",
          "first_token_ms",
          "cost_usd",
          "cost_per_question_usd",
          "tokens",
          "daily",
          "steps_ms"
        ],
        "title": "Overview"
      },
...
      "Percentiles": {
        "properties": {
          "p50": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50"
          },
          "p95": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95"
          }
        },
        "type": "object",
        "required": [
          "p50",
          "p95"
        ],
        "title": "Percentiles"
      },
      "QueryLogSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "condensed_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condensed Question"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "timings_ms": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Timings Ms"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "question",
          "condensed_question",
          "outcome",
          "timings_ms",
          "model",
          "cost_usd",
          "input_tokens",
          "output_tokens",
          "created_at"
        ],
        "title": "QueryLogSummary"
      },
      "QueryTrace": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "condensed_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condensed Question"
          },
          "allowed_collections": {
            "items": {
              "$ref": "#/components/schemas/CollectionRef"
            },
            "type": "array",
            "title": "Allowed Collections"
          },
          "retrieved": {
            "items": {
              "$ref": "#/components/schemas/RetrievedCandidate"
            },
            "type": "array",
            "title": "Retrieved"
          },
          "timings_ms": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Timings Ms"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          },
          "message_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Status"
          },
          "message_meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Message Meta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "conversation_id",
          "question",
          "condensed_question",
          "allowed_collections",
          "retrieved",
          "timings_ms",
          "model",
          "prompt_version",
          "input_tokens",
          "output_tokens",
          "cost_usd",
          "outcome",
          "prompt",
          "created_at",
          "message_id",
          "answer",
          "citations",
          "message_status",
          "message_meta"
        ],
        "title": "QueryTrace"
      },
      "RetrievedCandidate": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "vector_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Rank"
          },
          "vector_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Score"
          },
          "lexical_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexical Rank"
          },
          "lexical_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexical Score"
          },
          "rrf": {
            "type": "number",
            "title": "Rrf"
          },
          "rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Score"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "title",
          "rrf"
        ],
        "title": "RetrievedCandidate"
      },
...
      "TokenTotals": {
        "properties": {
          "input": {
            "type": "integer",
            "title": "Input"
          },
          "output": {
            "type": "integer",
            "title": "Output"
          }
        },
        "type": "object",
        "required": [
          "input",
          "output"
        ],
        "title": "TokenTotals"
      },
      "UnansweredCluster": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "users": {
            "type": "integer",
            "title": "Users"
          },
          "last_seen": {
            "type": "string",
            "title": "Last Seen"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ClusterQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "count",
          "users",
          "last_seen",
          "questions"
        ],
        "title": "UnansweredCluster"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}